+++ cli/deploy/mod.ts
@@ -1,10 +1,11 @@
+import { parseArkiveManifest } from '../../mod.ts'
 import { join, wait } from '../deps.ts'
 import { cleanup } from './cleanup.ts'
 import { pkg } from './pkg.ts'
 import { upload } from './upload.ts'
 
 export const action = async (
-	options: { public?: true; major?: true, env?: string },
+	options: { public?: true; major?: true; env?: string },
 	directory: string,
 ) => {
 	const dev = options.env?.toLowerCase() === 'dev'
@@ -33,6 +34,11 @@ export const action = async (
 		const { name: arkiveName } = manifest
 		if (!arkiveName) {
 			throw new Error(`Manifest must have a name property.`)
+		}
+
+		const { problems } = parseArkiveManifest.manifest(manifest)
+		if (problems) {
+			throw new Error(`Invalid manifest: ${problems}`)
 		}
 
 		spinner.text = 'Uploading package...'